+++ includes/class-backups-exception.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Backups exception class.
+ *
+ * @package EasyWPMigration
+ */
+
+defined( 'ABSPATH' ) || exit;
+
+/**
+ * Class EWPM_Backups_Exception
+ */
+class EWPM_Backups_Exception extends \Exception {
+}